refactor(hooks): extract custom data hooks from 7 page components #160
Separate data/state layer from view layer to prepare for real API integration.
Components now consume hooks for state management while keeping UI-only state local.

New hooks (src/lib/hooks/):
- useAnalyticsData — KPI, audit log filtering, export
- useSBOMData — SBOM CRUD, upload processing, vulnerability status
- useDeviceInventory — device filtering, sorting, pagination, CSV export
- useIncidentManagement — incident CRUD, device isolation, playbook progress
- useServiceOrders — order CRUD, kanban grouping, filtering, CSV export
- useComplianceManagement — compliance items, vulnerabilities, remediation
- useAuditLog — audit filtering, sorting, pagination, date validation
- useFirmwareDeployment — firmware lifecycle, SoD approval workflow
- useVulnerabilityTracker — vulnerability CRUD, severity filtering

New mock data (src/lib/mock-data/):
- analytics-data.ts, inventory-data.ts, service-order-data.ts, compliance-data.ts
