+++ src/app/components/sbom.tsx
@@ -1,24 +1,11 @@
-import { useState, useCallback } from "react";
+import { useState } from "react";
 import { FileBox } from "lucide-react";
-import { toast } from "sonner";
 import { cn } from "../../lib/utils";
 import { useAuth } from "../../lib/use-auth";
 import { getPrimaryRole } from "../../lib/rbac";
-import type {
-  Tab,
-  SBOMFormat,
-  SBOMStatus,
-  SBOM,
-  RemediationStatus,
-  ComponentVulnerability,
-} from "./sbom/sbom-types";
-import {
-  TABS,
-  MOCK_SBOMS,
-  MOCK_FIRMWARE,
-  MOCK_COMPONENTS,
-  MOCK_VULNERABILITIES,
-} from "./sbom/sbom-constants";
+import { useSBOMData } from "../../lib/hooks/use-sbom-data";
+import type { Tab } from "./sbom/sbom-types";
+import { TABS, MOCK_COMPONENTS } from "./sbom/sbom-constants";
 import { SBOMManagementTab } from "./sbom/sbom-management-tab";
 import { ComponentExplorerTab } from "./sbom/component-explorer-tab";
 import { CVEDashboardTab } from "./sbom/cve-dashboard-tab";
@@ -32,87 +19,19 @@ export function SBOMPage() {
   const { groups } = useAuth();
   const role = getPrimaryRole(groups);
   const [activeTab, setActiveTab] = useState<Tab>("management");
-  const [sboms, setSboms] = useState<SBOM[]>(MOCK_SBOMS);
-  const [vulnerabilities, setVulnerabilities] =
-    useState<ComponentVulnerability[]>(MOCK_VULNERABILITIES);
-  const [sbomFilter, setSbomFilter] = useState<string | null>(null);
+  const {
+    sboms,
+    vulnerabilities,
+    sbomFilter,
+    setSbomFilter,
+    handleUpload,
+    handleUpdateVulnStatus,
+  } = useSBOMData();
 
-  const handleUpload = useCallback((firmwareId: string, format: SBOMFormat, fileName: string) => {
-    const fw = MOCK_FIRMWARE.find((f) => f.id === firmwareId);
-    if (!fw) return;
-
-    const newSbom: SBOM = {
-      id: `sbom-${Date.now()}`,
-      firmwareId,
-      firmwareName: fw.name,
-      firmwareVersion: fw.version,
-      format,
-      specVersion: format === "CycloneDX" ? "1.5" : "2.3",
-      componentCount: 0,
-      vulnerabilityCount: 0,
-      licenseCount: 0,
-      criticalVulnCount: 0,
-      highVulnCount: 0,
-      mediumVulnCount: 0,
-      lowVulnCount: 0,
-      uploadedBy: "Current User",
-      uploadedDate: new Date().toISOString(),
-      status: "Processing",
-    };
-
-    setSboms((prev) => [newSbom, ...prev]);
-    toast.success(`SBOM uploaded for ${fw.name} v${fw.version}`, {
-      description: `${fileName} is being processed...`,
-    });
-
-    // Simulate processing completion after 3s
-    setTimeout(() => {
-      setSboms((prev) =>
-        prev.map((s) =>
-          s.id === newSbom.id
-            ? {
-                ...s,
-                status: "Complete" as SBOMStatus,
-                componentCount: Math.floor(Math.random() * 100) + 50,
-                vulnerabilityCount: Math.floor(Math.random() * 8),
-                licenseCount: Math.floor(Math.random() * 10) + 3,
-                criticalVulnCount: Math.floor(Math.random() * 2),
-                highVulnCount: Math.floor(Math.random() * 3),
-                mediumVulnCount: Math.floor(Math.random() * 3),
-                lowVulnCount: Math.floor(Math.random() * 2),
-              }
-            : s,
-        ),
-      );
-      toast.success("SBOM processing complete", {
-        description: `${fw.name} v${fw.version} SBOM has been parsed successfully`,
-      });
-    }, 3000);
-  }, []);
-
-  const handleViewDetails = useCallback((sbomId: string) => {
+  const handleViewDetails = (sbomId: string) => {
     setSbomFilter(sbomId);
     setActiveTab("components");
-  }, []);
-
-  const handleUpdateVulnStatus = useCallback(
-    (vulnId: string, newStatus: RemediationStatus, notes: string) => {
-      setVulnerabilities((prev) =>
-        prev.map((v) =>
-          v.id === vulnId
-            ? {
-                ...v,
-                remediationStatus: newStatus,
-                remediationNotes: notes,
-                resolvedDate: newStatus === "Resolved" ? new Date().toISOString() : v.resolvedDate,
-              }
-            : v,
-        ),
-      );
-      toast.success(`Vulnerability status updated to ${newStatus}`);
-    },
-    [],
-  );
+  };
 
   return (
     <div>
