+++ src/lib/hooks/use-vulnerability-tracker.ts
@@ -1,0 +1,79 @@
+import { useState, useMemo, useCallback } from "react";
+import { toast } from "sonner";
+import type {
+  VulnerabilityEntry,
+  VulnSeverity,
+  RemediationStatus,
+  AuditAction,
+} from "../../app/components/deployment/deployment-types";
+import {
+  INITIAL_VULNERABILITIES,
+  VULN_PAGE_SIZE,
+} from "../../app/components/deployment/deployment-constants";
+
+export function useVulnerabilityTracker(
+  addAuditEntry: (action: AuditAction, resourceType: string, resourceId: string) => void,
+) {
+  const [vulnerabilities, setVulnerabilities] =
+    useState<VulnerabilityEntry[]>(INITIAL_VULNERABILITIES);
+  const [vulnSeverityFilter, setVulnSeverityFilter] = useState<VulnSeverity | "All">("All");
+  const [vulnPage, setVulnPage] = useState(1);
+
+  const filteredVulnerabilities = useMemo(() => {
+    let items = vulnerabilities;
+    if (vulnSeverityFilter !== "All") {
+      items = items.filter((v) => v.severity === vulnSeverityFilter);
+    }
+    const severityOrder: Record<VulnSeverity, number> = { Critical: 0, High: 1, Medium: 2, Low: 3 };
+    items = [...items].sort((a, b) => severityOrder[a.severity] - severityOrder[b.severity]);
+    return items;
+  }, [vulnerabilities, vulnSeverityFilter]);
+
+  const totalVulnPages = Math.max(1, Math.ceil(filteredVulnerabilities.length / VULN_PAGE_SIZE));
+  const paginatedVulnerabilities = useMemo(() => {
+    const start = (vulnPage - 1) * VULN_PAGE_SIZE;
+    return filteredVulnerabilities.slice(start, start + VULN_PAGE_SIZE);
+  }, [filteredVulnerabilities, vulnPage]);
+
+  const handleCreateVulnerability = useCallback(
+    (data: Omit<VulnerabilityEntry, "id" | "resolvedDate">) => {
+      const newVuln: VulnerabilityEntry = {
+        ...data,
+        id: `vuln-${Date.now()}`,
+        resolvedDate: null,
+      };
+      setVulnerabilities((prev) => [newVuln, ...prev]);
+      addAuditEntry("Created", "Vulnerability", `VULN#${newVuln.id}`);
+      toast.success("Vulnerability record created");
+    },
+    [addAuditEntry],
+  );
+
+  const handleRemediationChange = useCallback(
+    (vulnId: string, newStatus: RemediationStatus) => {
+      const resolvedDate =
+        newStatus === "Resolved" ? (new Date().toISOString().split("T")[0] ?? null) : null;
+      setVulnerabilities((prev) =>
+        prev.map((v) =>
+          v.id === vulnId ? { ...v, remediationStatus: newStatus, resolvedDate } : v,
+        ),
+      );
+      addAuditEntry("Modified", "Vulnerability", `VULN#${vulnId}`);
+      toast.success(`Vulnerability status updated to ${newStatus}`);
+    },
+    [addAuditEntry],
+  );
+
+  return {
+    vulnerabilities,
+    filteredVulnerabilities,
+    paginatedVulnerabilities,
+    totalVulnPages,
+    vulnSeverityFilter,
+    setVulnSeverityFilter,
+    vulnPage,
+    setVulnPage,
+    handleCreateVulnerability,
+    handleRemediationChange,
+  };
+}
