+++ src/lib/mock-data/inventory-data.ts
@@ -1,0 +1,221 @@
+import { DeviceStatus } from "../types";
+
+export interface MockDevice {
+  id: string;
+  name: string;
+  serial: string;
+  model: string;
+  status: DeviceStatus;
+  location: string;
+  health: number;
+  firmware: string;
+  lastSeen: string;
+  lat?: number;
+  lng?: number;
+}
+
+export const MOCK_DEVICES: MockDevice[] = [
+  {
+    id: "d1",
+    name: "INV-3200A",
+    serial: "SN-4821",
+    model: "INV-3200",
+    status: DeviceStatus.Online,
+    location: "Denver, CO",
+    health: 98,
+    firmware: "v4.0.0",
+    lastSeen: "2m ago",
+    lat: 39.74,
+    lng: -104.99,
+  },
+  {
+    id: "d2",
+    name: "INV-3200B",
+    serial: "SN-4822",
+    model: "INV-3200",
+    status: DeviceStatus.Online,
+    location: "Houston, TX",
+    health: 95,
+    firmware: "v4.0.0",
+    lastSeen: "5m ago",
+    lat: 29.76,
+    lng: -95.37,
+  },
+  {
+    id: "d3",
+    name: "INV-3100C",
+    serial: "SN-3901",
+    model: "INV-3100",
+    status: DeviceStatus.Maintenance,
+    location: "Chicago, IL",
+    health: 72,
+    firmware: "v3.2.1",
+    lastSeen: "1h ago",
+    lat: 41.88,
+    lng: -87.63,
+  },
+  {
+    id: "d4",
+    name: "INV-3200D",
+    serial: "SN-4892",
+    model: "INV-3200",
+    status: DeviceStatus.Offline,
+    location: "Denver, CO",
+    health: 0,
+    firmware: "v4.0.0",
+    lastSeen: "30m ago",
+    lat: 39.74,
+    lng: -104.99,
+  },
+  {
+    id: "d5",
+    name: "INV-3100E",
+    serial: "SN-3455",
+    model: "INV-3100",
+    status: DeviceStatus.Online,
+    location: "New York, NY",
+    health: 91,
+    firmware: "v3.2.1",
+    lastSeen: "1m ago",
+    lat: 40.71,
+    lng: -74.01,
+  },
+  {
+    id: "d6",
+    name: "INV-3200F",
+    serial: "SN-5001",
+    model: "INV-3200",
+    status: DeviceStatus.Online,
+    location: "Dallas, TX",
+    health: 99,
+    firmware: "v4.0.0",
+    lastSeen: "3m ago",
+    lat: 32.78,
+    lng: -96.8,
+  },
+  {
+    id: "d7",
+    name: "INV-3100G",
+    serial: "SN-3102",
+    model: "INV-3100",
+    status: DeviceStatus.Decommissioned,
+    location: "Phoenix, AZ",
+    health: 0,
+    firmware: "v3.1.0",
+    lastSeen: "7d ago",
+    lat: 33.45,
+    lng: -112.07,
+  },
+  {
+    id: "d8",
+    name: "INV-3200H",
+    serial: "SN-5102",
+    model: "INV-3200",
+    status: DeviceStatus.Online,
+    location: "Shanghai, CN",
+    health: 97,
+    firmware: "v4.0.0",
+    lastSeen: "1m ago",
+    lat: 31.23,
+    lng: 121.47,
+  },
+  {
+    id: "d9",
+    name: "INV-3100J",
+    serial: "SN-3210",
+    model: "INV-3100",
+    status: DeviceStatus.Maintenance,
+    location: "Munich, DE",
+    health: 65,
+    firmware: "v3.2.1",
+    lastSeen: "2h ago",
+    lat: 48.14,
+    lng: 11.58,
+  },
+  {
+    id: "d10",
+    name: "INV-3200K",
+    serial: "SN-5201",
+    model: "INV-3200",
+    status: DeviceStatus.Online,
+    location: "Singapore, SG",
+    health: 94,
+    firmware: "v4.0.0",
+    lastSeen: "4m ago",
+    lat: 1.35,
+    lng: 103.82,
+  },
+  {
+    id: "d11",
+    name: "INV-3100L",
+    serial: "SN-3301",
+    model: "INV-3100",
+    status: DeviceStatus.Offline,
+    location: "Sao Paulo, BR",
+    health: 0,
+    firmware: "v3.1.0",
+    lastSeen: "45m ago",
+    lat: -23.55,
+    lng: -46.63,
+  },
+  {
+    id: "d12",
+    name: "INV-3200M",
+    serial: "SN-5301",
+    model: "INV-3200",
+    status: DeviceStatus.Online,
+    location: "Denver, CO",
+    health: 88,
+    firmware: "v4.0.0",
+    lastSeen: "8m ago",
+    lat: 39.74,
+    lng: -104.99,
+  },
+  {
+    id: "d13",
+    name: "INV-3200N",
+    serial: "SN-5401",
+    model: "INV-3200",
+    status: DeviceStatus.Online,
+    location: "Sydney",
+    health: 92,
+    firmware: "v4.0.0",
+    lastSeen: "6m ago",
+    // No lat/lng — Story 9.5 fallback to location name lookup
+  },
+  {
+    id: "d14",
+    name: "INV-3100P",
+    serial: "SN-3401",
+    model: "INV-3100",
+    status: DeviceStatus.Offline,
+    location: "Tokyo",
+    health: 0,
+    firmware: "v3.2.1",
+    lastSeen: "1h ago",
+    lat: 0,
+    lng: 0,
+    // lat/lng are 0,0 — Story 9.5 AC4: treated as missing, fallback to lookup
+  },
+  {
+    id: "d15",
+    name: "INV-3100Q",
+    serial: "SN-3501",
+    model: "INV-3100",
+    status: DeviceStatus.Maintenance,
+    location: "Unknown Base",
+    health: 55,
+    firmware: "v3.1.0",
+    lastSeen: "3h ago",
+    // No coordinates, location not in lookup — Story 9.5 AC3: excluded from map
+  },
+];
+
+export const ALL_STATUSES = [
+  DeviceStatus.Online,
+  DeviceStatus.Offline,
+  DeviceStatus.Maintenance,
+  DeviceStatus.Decommissioned,
+];
+export const ALL_LOCATIONS = [...new Set(MOCK_DEVICES.map((d) => d.location))].sort();
+export const ALL_MODELS = [...new Set(MOCK_DEVICES.map((d) => d.model))].sort();
